refactor: WIP - Tight boundaries between generic spatial skeleton and specific implementation
