+++ src/datasource/catmaid/backend.ts
@@ -97,7 +97,7 @@ export class CatmaidSpatiallyIndexedSkeletonSourceBackend extends WithParameters
     // Pack only segment IDs into vertexAttributes (positions are in vertexPositions)
     chunk.vertexAttributes = [packed.segmentIds];
     chunk.nodeIds = packed.nodeIds;
-    chunk.nodeRevisionTokens = packed.revisionTokens;
+    chunk.nodeSourceStates = packed.sourceStates;
   }
 }
