+++ src/datasource/catmaid/skeleton_packing.ts
@@ -21,7 +21,7 @@ interface PackedCatmaidSkeletonData {
   segmentIds: Uint32Array;
   indices: Uint32Array;
   nodeIds: Int32Array;
-  revisionTokens: Array<string | undefined>;
+  sourceStates: unknown[];
 }
 
 export function packCatmaidSkeletonNodes(
@@ -31,7 +31,7 @@ export function packCatmaidSkeletonNodes(
   const vertexPositions = new Float32Array(numVertices * 3);
   const segmentIds = new Uint32Array(numVertices);
   const nodeIds = new Int32Array(numVertices);
-  const revisionTokens = new Array<string | undefined>(numVertices);
+  const sourceStates = new Array<unknown>(numVertices);
   const indices: number[] = [];
   const nodeMap = new Map<number, number>();
 
@@ -43,7 +43,7 @@ export function packCatmaidSkeletonNodes(
     vertexPositions[i * 3 + 1] = node.position[1];
     vertexPositions[i * 3 + 2] = node.position[2];
     segmentIds[i] = node.segmentId;
-    revisionTokens[i] = node.revisionToken;
+    sourceStates[i] = node.sourceState;
   }
 
   for (let i = 0; i < numVertices; ++i) {
@@ -60,6 +60,6 @@ export function packCatmaidSkeletonNodes(
     segmentIds,
     indices: new Uint32Array(indices),
     nodeIds,
-    revisionTokens,
+    sourceStates,
   };
 }
