+++ src/rendered_data_panel.ts
@@ -65,7 +65,7 @@ interface SpatialSkeletonSelectableLayer {
     options?: {
       segmentId?: number;
       position?: ArrayLike<number>;
-      revisionToken?: string;
+      sourceState?: unknown;
     },
   ) => void;
   clearSpatialSkeletonNodeSelection: (
@@ -539,7 +539,7 @@ export abstract class RenderedDataPanel extends RenderedPanel {
             ? pickedSegmentId
             : undefined,
         position: pickedSpatialSkeleton?.position ?? mouseState.position,
-        revisionToken: pickedSpatialSkeleton?.revisionToken,
+        sourceState: pickedSpatialSkeleton?.sourceState,
       };
     };
 
@@ -570,7 +570,7 @@ export abstract class RenderedDataPanel extends RenderedPanel {
           {
             segmentId: pickedSelection.segmentId,
             position: pickedSelection.position,
-            revisionToken: pickedSelection.revisionToken,
+            sourceState: pickedSelection.sourceState,
           },
         );
         return;
